// src/api/graphql-types/__generated__/graphql-types.ts
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export enum LrtLine {
  BlueLine = 'BLUE_LINE',
  RedLine = 'RED_LINE'
}

export type LrtStation = {
  __typename?: 'LRTStation';
  lines: Array<LrtLine>;
  name: Scalars['String']['output'];
};

export type Query = {
  __typename?: 'Query';
  calgaryLRTStations: Array<LrtStation>;
  calgaryRivers: Array<River>;
};

export type River = {
  __typename?: 'River';
  lengthInKilometers: Scalars['Int']['output'];
  name: Scalars['String']['output'];
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  Int: ResolverTypeWrapper<Scalars['Int']['output']>;
  LRTLine: LrtLine;
  LRTStation: ResolverTypeWrapper<LrtStation>;
  Query: ResolverTypeWrapper<{}>;
  River: ResolverTypeWrapper<River>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean']['output'];
  Int: Scalars['Int']['output'];
  LRTStation: LrtStation;
  Query: {};
  River: River;
  String: Scalars['String']['output'];
};

export type LrtStationResolvers<ContextType = any, ParentType extends ResolversParentTypes['LRTStation'] = ResolversParentTypes['LRTStation']> = {
  lines?: Resolver<Array<ResolversTypes['LRTLine']>, ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  calgaryLRTStations?: Resolver<Array<ResolversTypes['LRTStation']>, ParentType, ContextType>;
  calgaryRivers?: Resolver<Array<ResolversTypes['River']>, ParentType, ContextType>;
};

export type RiverResolvers<ContextType = any, ParentType extends ResolversParentTypes['River'] = ResolversParentTypes['River']> = {
  lengthInKilometers?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type Resolvers<ContextType = any> = {
  LRTStation?: LrtStationResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  River?: RiverResolvers<ContextType>;
};